add structure to torchtitan files
