+++ docs/converging.md
@@ -35,11 +35,11 @@ If the technique is not a parallelism
 
 ## Example
 
-This is a series of loss-converging tests covering both parallelisms and training optimizations.
+This is a series of loss-converging tests on Llama 3.1, covering both parallelisms and training optimizations.
 Results are obtained on 2025/01/21, with the latest `torch`, `torchao`, and `torchtitan`.
 
 ### Setup
-- Base config: [train_configs/llama3_8b.toml](../train_configs/llama3_8b.toml)
+- Base config: [torchtitan/models/llama/train_configs/llama3_8b.toml](../torchtitan/models/llama/train_configs/llama3_8b.toml)
 - `training.batch_size = 4`, which is a minimum for Pipeline Parallel with `pipeline_parallel_degree = 2` and `pipeline_parallel_schedule = "Interleaved1F1B"`
 - `training.data_parallel_shard_degree = 8`, resulting in global batch size 32
 - `training.steps = 3000`, `training.warmup_steps = 600`
