+++ scripts/generate/run_llama_generate.sh
@@ -12,7 +12,7 @@ set -e
 # LOG_RANK=0,1 NGPU=4 ./run_llama_generate.sh
 NGPU=${NGPU:-"1"}
 LOG_RANK=${LOG_RANK:-0}
-CONFIG_FILE=${CONFIG_FILE:-"./train_configs/debug_model.toml"}
+CONFIG_FILE=${CONFIG_FILE:-"./torchtitan/models/llama/train_configs/debug_model.toml"}
 CHECKPOINT_DIR=${CHECKPOINT_DIR:-"./outputs/checkpoint/"}
 PROMPT=${PROMPT:-""}
