+++ tests/integration_tests.py
@@ -522,7 +522,9 @@ def run_tests(args):
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("output_dir")
-    parser.add_argument("--config_dir", default="./train_configs")
+    parser.add_argument(
+        "--config_dir", default="./torchtitan/models/llama/train_configs"
+    )
     parser.add_argument(
         "--test",
         default="all",
