+++ torchtitan/__init__.py
@@ -7,7 +7,7 @@
 # Copyright (c) Meta Platforms, Inc. All Rights Reserved.
 
 # Import to register Float8Converter.
-import torchtitan.float8  # noqa: F401
+import torchtitan.components.float8  # noqa: F401
 
 # Import the built-in models here so that the corresponding register_model_spec()
 # will be called.
