+++ torchtitan/datasets/hf_datasets.py
@@ -14,10 +14,10 @@ from datasets.distributed import split_dataset_by_node
 from torch.distributed.checkpoint.stateful import Stateful
 from torch.utils.data import IterableDataset
 
+from torchtitan.components.dataloader import ParallelAwareDataloader
+from torchtitan.components.tokenizer import Tokenizer
 from torchtitan.config_manager import JobConfig
-from torchtitan.dataloader import ParallelAwareDataloader
-from torchtitan.datasets.tokenizer import Tokenizer
-from torchtitan.logging import logger
+from torchtitan.tools.logging import logger
 
 
 def _load_c4_dataset(dataset_path: str):
