+++ torchtitan/distributed/parallel_dims.py
@@ -9,7 +9,7 @@ from functools import cached_property
 
 from torch.distributed.device_mesh import init_device_mesh
 
-from torchtitan.logging import logger
+from torchtitan.tools.logging import logger
 
 
 __all__ = ["ParallelDims"]
