+++ torchtitan/distributed/utils.py
@@ -5,13 +5,8 @@
 # LICENSE file in the root directory of this source tree.
 
 import contextlib
-import gc
-import importlib
 import math
 import os
-import subprocess
-import sys
-from dataclasses import dataclass
 from datetime import timedelta
 from typing import Generator, Iterable, List, Optional, Set, Union
 
@@ -19,25 +14,14 @@ import torch
 import torch.distributed._functional_collectives as funcol
 import torch.distributed.distributed_c10d as c10d
 from torch import distributed as dist
-from torch._utils import _get_available_device_type, _get_device_module
 from torch.distributed.device_mesh import DeviceMesh
 from torch.distributed.tensor import DTensor
 
-from torchtitan.logging import logger
+from torchtitan.tools.logging import logger
+from torchtitan.tools.utils import device_module, device_type
 
 
-def get_device_info():
-    device_type = _get_available_device_type()
-    if device_type is None:
-        device_type = "cuda"  # default device_type: cuda
-    device_module = _get_device_module(device_type)  # default device_module:torch.cuda
-    return device_type, device_module
-
-
-device_type, device_module = get_device_info()
-
-
-def dist_reduce(x: torch.Tensor, reduceOp: str, mesh: DeviceMesh) -> float:
+def _dist_reduce(x: torch.Tensor, reduceOp: str, mesh: DeviceMesh) -> float:
     if isinstance(x, DTensor):
         # functional collectives do not support DTensor inputs
         x = x.full_tensor()
@@ -46,19 +30,11 @@ def dist_reduce(x: torch.Tensor, reduceOp: str, mesh: DeviceMesh) -> float:
 
 
 def dist_max(x: torch.Tensor, mesh: DeviceMesh) -> float:
-    return dist_reduce(x, reduceOp=c10d.ReduceOp.MAX.name, mesh=mesh)
+    return _dist_reduce(x, reduceOp=c10d.ReduceOp.MAX.name, mesh=mesh)
 
 
 def dist_mean(x: torch.Tensor, mesh: DeviceMesh) -> float:
-    return dist_reduce(x, reduceOp=c10d.ReduceOp.AVG.name, mesh=mesh)
-
-
-def _warn_overwrite_env(env, val):
-    if env in os.environ:
-        logger.warning(
-            f"ENV[{env}] = {os.environ[env]} will be overridden to {val} based on job config"
-        )
-    os.environ[env] = val
+    return _dist_reduce(x, reduceOp=c10d.ReduceOp.AVG.name, mesh=mesh)
 
 
 def set_determinism(
@@ -130,58 +106,6 @@ def set_determinism(
         torch.distributed.tensor._random.manual_seed(seed, spmd_mesh)
 
 
-def set_pg_timeouts(timeout, world_mesh):
-    """
-    Sets the timeout for all PGs in the provided mesh, and the default (world) group.
-
-    Note: synchronizes via a barrier, before changing the timeouts. This is important, because
-    otherwise you may face a race where the slow rank has not reached the timeout reduction point
-    yet due to slow operations permitted under the old timeout value, but other faster ranks may
-    start issuing collectives under the new shorter timeout and then immediately timeout.
-    """
-    logger.info(
-        f"Synchronizing and adjusting timeout for all ProcessGroups to {timeout}"
-    )
-    # Ensure that all the ranks have reached the point of setting the new timeout-
-    # otherwise, some ranks may issue collectives with the new/shorter timeout and
-    # those may time out, before other ranks have finished with initialization done
-    # under the old/slow timeout.
-    torch.distributed.barrier(device_ids=[device_module.current_device()])
-    device_module.synchronize()
-
-    groups = [world_mesh.get_group(mesh_dim) for mesh_dim in range(world_mesh.ndim)]
-
-    # None represents the 'default' PG, not part of the mesh
-    groups.append(None)
-    for group in groups:
-        torch.distributed.distributed_c10d._set_pg_timeout(timeout, group)
-
-
-# used to avoid stragglers in garbage collection
-class GarbageCollection:
-    def __init__(self, gc_freq=1000):
-        assert gc_freq > 0, "gc_freq must be a positive integer"
-        self.gc_freq = gc_freq
-        gc.disable()
-        self.collect("Initial GC collection.")
-
-    def run(self, step_count):
-        if step_count > 1 and step_count % self.gc_freq == 0:
-            self.collect("Peforming periodical GC collection.")
-
-    @staticmethod
-    def collect(reason: str):
-        logger.info(reason)
-        gc.collect(1)
-
-
-TRACE_BUFFER_SIZE = "TORCH_NCCL_TRACE_BUFFER_SIZE"
-TRACE_FILE = "TORCH_NCCL_DEBUG_INFO_TEMP_FILE"
-DUMP_ON_TIMEOUT = "TORCH_NCCL_DUMP_ON_TIMEOUT"
-ASYNC_ERROR_HANDLING = "TORCH_NCCL_ASYNC_ERROR_HANDLING"
-SKIP_CLEANUP = "3"
-
-
 def create_context_parallel_ctx(
     cp_mesh: DeviceMesh,
     cp_buffers: List[torch.Tensor],
@@ -236,16 +160,30 @@ def get_train_context(enable_loss_parallel: bool, enable_compiled_autograd: bool
     return context
 
 
-def _get_distributed_backend(job_config):
-    backend = "nccl"
-    if device_type in torch.distributed.Backend.default_device_backend_map.keys():
-        backend = torch.distributed.Backend.default_device_backend_map.get(device_type)
-    if job_config.training.enable_cpu_offload:
-        backend = f"{device_type}:{backend},cpu:gloo"
-    return backend
-
-
 def init_distributed(job_config):
+    def _warn_overwrite_env(env, val):
+        if env in os.environ:
+            logger.warning(
+                f"ENV[{env}] = {os.environ[env]} will be overridden to {val} based on job config"
+            )
+        os.environ[env] = val
+
+    def _get_distributed_backend(job_config):
+        backend = "nccl"
+        if device_type in torch.distributed.Backend.default_device_backend_map.keys():
+            backend = torch.distributed.Backend.default_device_backend_map.get(
+                device_type
+            )
+        if job_config.training.enable_cpu_offload:
+            backend = f"{device_type}:{backend},cpu:gloo"
+        return backend
+
+    TRACE_BUFFER_SIZE = "TORCH_NCCL_TRACE_BUFFER_SIZE"
+    TRACE_FILE = "TORCH_NCCL_DEBUG_INFO_TEMP_FILE"
+    DUMP_ON_TIMEOUT = "TORCH_NCCL_DUMP_ON_TIMEOUT"
+    ASYNC_ERROR_HANDLING = "TORCH_NCCL_ASYNC_ERROR_HANDLING"
+    SKIP_CLEANUP = "3"
+
     # FlightRecorder is incompatible with =1 mode where watchdog aborts work, must use =3 (skipcleanup)
     # to get flight recorder dumps. See https://github.com/pytorch/pytorch/issues/121055
     # This could be done only when flight recorder is enabled, but its nice to be consistent to avoid subtle
@@ -272,90 +210,31 @@ def init_distributed(job_config):
     )
 
 
-def get_num_params(model: torch.nn.Module, exclude_embedding: bool = False) -> int:
-    num_params = sum(p.numel() for p in model.parameters())
-    if exclude_embedding:
-        num_params -= sum(p.numel() for p in model.tok_embeddings.parameters())
-    return num_params
+def set_pg_timeouts(timeout, world_mesh):
+    """
+    Sets the timeout for all PGs in the provided mesh, and the default (world) group.
 
-
-def get_num_flop_per_token(num_params: int, model_config, seq_len) -> int:
-    l, h, q, t = (
-        model_config.n_layers,
-        model_config.n_heads,
-        model_config.dim // model_config.n_heads,
-        seq_len,
+    Note: synchronizes via a barrier, before changing the timeouts. This is important, because
+    otherwise you may face a race where the slow rank has not reached the timeout reduction point
+    yet due to slow operations permitted under the old timeout value, but other faster ranks may
+    start issuing collectives under the new shorter timeout and then immediately timeout.
+    """
+    logger.info(
+        f"Synchronizing and adjusting timeout for all ProcessGroups to {timeout}"
     )
-    # Reasoning behind the factor of 12 for the self-attention part of the formula:
-    # 1. each self-attention has 2 matmul in the forward and 4 in the backward (6)
-    # 2. the flash attention does 1 more matmul recomputation in the backward
-    #    but recomputation should not be counted in calculating MFU           (+0)
-    # 3. each matmul performs 1 multiplication and 1 addition                 (*2)
-    # 4. we follow the convention and do not account for sparsity in causal attention
-    flop_per_token = 6 * num_params + 12 * l * h * q * t
+    # Ensure that all the ranks have reached the point of setting the new timeout-
+    # otherwise, some ranks may issue collectives with the new/shorter timeout and
+    # those may time out, before other ranks have finished with initialization done
+    # under the old/slow timeout.
+    torch.distributed.barrier(device_ids=[device_module.current_device()])
+    device_module.synchronize()
 
-    return flop_per_token
+    groups = [world_mesh.get_group(mesh_dim) for mesh_dim in range(world_mesh.ndim)]
 
-
-# hardcoded BF16 type peak flops for NVIDIA A100, H100, and H200 GPU
-def get_peak_flops(device_name: str) -> int:
-    try:
-        # Run the lspci command and capture the output
-        result = subprocess.run(["lspci"], stdout=subprocess.PIPE, text=True)
-        # Filter the output for lines containing both "NVIDIA" and "H100"
-        filtered_lines = [
-            line
-            for line in result.stdout.splitlines()
-            if "NVIDIA" in line and "H100" in line
-        ]
-        # Join all filtered lines into a single string
-        device_name = " ".join(filtered_lines) or device_name
-    except FileNotFoundError as e:
-        logger.warning(f"Error running lspci: {e}, fallback to use device_name")
-    if "A100" in device_name:
-        # data from https://www.nvidia.com/en-us/data-center/a100/
-        return 312e12
-    elif "H100" in device_name:
-        # data from https://www.nvidia.com/en-us/data-center/h100/
-        # NOTE: Specifications are one-half lower without sparsity.
-        if "NVL" in device_name:
-            return 835e12
-        elif "PCIe" in device_name:
-            return 756e12
-        else:  # for H100 SXM and other variants
-            return 989e12
-    elif "H200" in device_name:
-        # data from https://www.nvidia.com/en-us/data-center/h200/
-        return 989e12
-    else:  # for other GPU types, assume A100
-        logger.warning(f"Peak flops undefined for: {device_name}, fallback to A100")
-        return 312e12
-
-
-@dataclass(frozen=True)
-class Color:
-    black = "\033[30m"
-    red = "\033[31m"
-    green = "\033[32m"
-    yellow = "\033[33m"
-    blue = "\033[34m"
-    magenta = "\033[35m"
-    cyan = "\033[36m"
-    white = "\033[37m"
-    reset = "\033[39m"
-
-
-@dataclass(frozen=True)
-class NoColor:
-    black = ""
-    red = ""
-    green = ""
-    yellow = ""
-    blue = ""
-    magenta = ""
-    cyan = ""
-    white = ""
-    reset = ""
+    # None represents the 'default' PG, not part of the mesh
+    groups.append(None)
+    for group in groups:
+        torch.distributed.distributed_c10d._set_pg_timeout(timeout, group)
 
 
 @torch.no_grad()
@@ -419,55 +298,3 @@ def clip_grad_norm_(
 
     torch.nn.utils.clip_grads_with_norm_(parameters, max_norm, total_norm, foreach)
     return total_norm
-
-
-def check_if_feature_in_pytorch(
-    feature_name: str,
-    pull_request: str,
-    min_nightly_version: Optional[str] = None,
-) -> None:
-    if "git" in torch.__version__:  # pytorch is built from source
-        # notify users to check if the pull request is included in their pytorch
-        logger.warning(
-            "detected that the pytorch is built from source. Please make sure the PR "
-            f"({pull_request_link}) is included in pytorch for correct {feature_name}."
-        )
-    elif min_nightly_version is not None and torch.__version__ < min_nightly_version:
-        logger.warning(
-            f"detected that the pytorch version {torch.__version__} is older than "
-            f"{min_nightly_version}. Please upgrade a newer version to include the "
-            f"change in ({pull_request_link}) for correct {feature_name}."
-        )
-
-
-def import_module_from_path(path: str):
-    path = os.path.expanduser(path)
-
-    # 1. Check if path is an existing file or directory path.
-    if os.path.exists(path):
-        if not os.path.isdir(path):
-            raise ImportError(f"Path '{path}' is not a directory.")
-        init_file = os.path.join(path, "__init__.py")
-        if os.path.isfile(init_file):
-            return _import_module_from_init(path)
-
-        raise ImportError(
-            f"Directory '{path}' is not a Python package because it does not "
-            "contain an __init__.py file."
-        )
-
-    # 2. If not a valid path, assume it's a dotted module name.
-    return importlib.import_module(path)
-
-
-def _import_module_from_init(path: str):
-    init_file = os.path.join(path, "__init__.py")
-    module_name = os.path.basename(path)
-    spec = importlib.util.spec_from_file_location(module_name, init_file)
-    if spec is None:
-        raise ImportError(f"Could not create spec from '{init_file}'")
-
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[module_name] = module
-    spec.loader.exec_module(module)
-    return module
