+++ torchtitan/models/llama/__init__.py
@@ -6,12 +6,12 @@
 #
 # Copyright (c) Meta Platforms, Inc. All Rights Reserved.
 
+from torchtitan.components.loss import cross_entropy_loss
+from torchtitan.components.optimizer import build_lr_schedulers, build_optimizers
 from torchtitan.datasets import build_hf_dataloader
 from torchtitan.datasets.tokenizer import TikTokenizer
-from torchtitan.loss import cross_entropy_loss
 from torchtitan.models.llama.model import Transformer, TransformerModelArgs
-from torchtitan.optimizer import build_lr_schedulers, build_optimizers
-from torchtitan.train_spec import register_train_spec, TrainSpec
+from torchtitan.protocols.train_spec import register_train_spec, TrainSpec
 
 from .parallelize_llama import parallelize_llama
 from .pipeline_llama import pipeline_llama
