+++ torchtitan/models/llama/pipeline_llama.py
@@ -20,13 +20,13 @@ from torch.distributed.pipelining.schedules import (
 )
 
 from torchtitan.config_manager import JobConfig
-from torchtitan.logging import logger
-from torchtitan.parallelisms import ParallelDims
-from torchtitan.parallelisms.pipeline import (
+from torchtitan.distributed import ParallelDims
+from torchtitan.distributed.pipeline import (
     build_pipeline_schedule,
     generate_split_points,
     stage_ids_this_rank,
 )
+from torchtitan.tools.logging import logger
 
 from .model import TransformerModelArgs
