+++ torchtitan/protocols/model_converter.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Protocol, Union
 import torch.nn as nn
 
 from torchtitan.config_manager import JobConfig
-from torchtitan.parallelisms import ParallelDims
+from torchtitan.distributed import ParallelDims
 
 
 class ModelConverter(Protocol):
