+++ torchtitan/tools/metrics.py
@@ -13,9 +13,9 @@ import torch
 from torch.utils.tensorboard import SummaryWriter
 
 from torchtitan.config_manager import JobConfig
-from torchtitan.logging import logger
-from torchtitan.parallelisms import ParallelDims
-from torchtitan.utils import device_module, device_type
+from torchtitan.distributed import ParallelDims
+from torchtitan.tools.logging import logger
+from torchtitan.tools.utils import device_module, device_type
 
 # named tuple for passing device memory stats for logging
 DeviceMemStats = namedtuple(
